+++ source-files/js/_afterLoad.js
@@ -66,6 +66,7 @@ if (!(window.customElements && document.body.attachShadow)) {
   document.getElementById('ircServerPanel').initializePlugin();
   document.getElementById('wallopsPanel').initializePlugin();
   document.getElementById('noticePanel').initializePlugin();
+  document.getElementById('manageChannelsPanel').initializePlugin();
   document.getElementById('managePmPanels').initializePlugin();
 
   /**