+++ source-files/web-components/notice-panel.js
@@ -263,7 +263,7 @@ window.customElements.define('notice-panel', class extends HTMLElement {
     // Normal button handler
     // -------------------------
     this.shadowRoot.getElementById('normalButtonId').addEventListener('click', () => {
-      this.shadowRoot.getElementById('panelMessageDisplayId').setAttribute('rows', '5');
+      this.shadowRoot.getElementById('panelMessageDisplayId').setAttribute('rows', '10');
     });
 
     // -------------------------------