+++ source-files/web-components/wallops-panel.js
@@ -184,7 +184,7 @@ window.customElements.define('wallops-panel', class extends HTMLElement {
     // Normal button handler
     // -------------------------
     this.shadowRoot.getElementById('normalButtonId').addEventListener('click', () => {
-      this.shadowRoot.getElementById('panelMessageDisplayId').setAttribute('rows', '5');
+      this.shadowRoot.getElementById('panelMessageDisplayId').setAttribute('rows', '10');
     });
 
     // -------------------------------